task2/requirements.txt
torch>=2.2.0
 torchvision>=0.17.0
 numpy>=1.24.3,<2.0
 pandas>=2.0.2
 scikit-learn>=1.3.0
 matplotlib>=3.7.1
 seaborn>=0.12.2
 opencv-python>=4.7.0.72
 pillow>=9.5.0
 tqdm>=4.65.0